move stock state signals into stock app
